requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
numpy>=1.21.0
pydantic>=2.0.0
matplotlib>=3.5.0
diffusers>=0.24.0
PyYAML>=6.0
tqdm>=4.65.0
scipy>=1.10.0 